chore: dealwith cr
