+++ packages/plugin-pha/package.json
@@ -17,7 +17,6 @@
   },
   "dependencies": {
     "@remix-run/router": "^1.0.3",
-    "build-scripts": "^2.0.1-0",
     "chalk": "^4.0.0",
     "consola": "^2.15.3",
     "humps": "^2.0.1",
@@ -25,6 +24,7 @@
   },
   "devDependencies": {
     "@ice/app": "^3.0.0",
+    "build-scripts": "^2.0.1-0",
     "esbuild": "^0.16.5",
     "webpack": "^5.73.0",
     "webpack-dev-server": "^4.9.2"
