+++ packages/plugin-pha/src/phaMiddleware.ts
@@ -4,6 +4,7 @@ import * as fs from 'fs';
 import type { ExpressRequestHandler } from 'webpack-dev-server';
 import consola from 'consola';
 import type { PluginData } from '@ice/app/esm/types';
+import { getCompilerConfig } from './constants';
 import { parseManifest, rewriteAppWorker, getAppWorkerUrl, getMultipleManifest, type ParseOptions } from './manifestHelpers.js';
 import { getAppWorkerContent, type Options } from './generateManifest.js';
 import type { Manifest } from './types.js';
@@ -46,36 +47,14 @@ const createPHAMiddleware = ({
         manifest = rewriteAppWorker(manifest);
         if (requestAppWorker) {
           const entry = path.join(rootDir, './.ice/appWorker.ts');
-          const plugins = getAllPlugin(['keepExports']) as PluginData[];
-          let keepExports = ['dataLoader'];
-          plugins.forEach(plugin => {
-            if (plugin.keepExports) {
-              keepExports = keepExports.concat(plugin.keepExports);
-            }
-          });
           sendResponse(
             res,
             await getAppWorkerContent(compiler, {
               entry: fs.existsSync(entry) ? entry : appWorkerPath,
               outfile: path.join(outputDir, 'app-worker.js'),
-            }, {
-              swc: {
-                keepExports,
-                keepPlatform: 'web',
-                getRoutePaths: () => {
-                  return ['src/pages'];
-                },
-              },
-              preBundle: false,
-              externalDependencies: false,
-              transformEnv: false,
-              enableEnv: true,
-              // Redirect import defineDataLoader from @ice/runtime to avoid build plugin side effect code.
-              redirectImports: [{
-                specifier: ['defineDataLoader'],
-                source: '@ice/runtime',
-              }],
-            }),
+            }, getCompilerConfig({
+              getAllPlugin,
+            })),
             'text/javascript',
           );
           return;
